deduplication of functions for scripts in utils.py
